+++ src/BIT.Data.Sync.Xpo.DeltaStore/XpoDelta.cs
@@ -5,52 +5,6 @@ using System.Linq;
 
 namespace BIT.Data.Sync.Xpo.DeltaStore
 {
-    public class XpoSyncStatus : XPObject, ISyncStatus
-    {
-        public XpoSyncStatus(Session session) : base(session)
-        { }
-
-        int lastTransactionLogProcessed;
-        string lastPushedDelta;
-        string lastProcessedDelta;
-        string identity;
-        Guid id;
-
-        [Key(autoGenerate:true)]
-        public Guid Id
-        {
-            get => id;
-            private set => SetPropertyValue(nameof(Id), ref id, value);
-        }
-
-        [Size(SizeAttribute.DefaultStringMappingFieldSize)]
-        public string Identity
-        {
-            get => identity;
-            set => SetPropertyValue(nameof(Identity), ref identity, value);
-        }
-
-        [Size(SizeAttribute.DefaultStringMappingFieldSize)]
-        public string LastProcessedDelta
-        {
-            get => lastProcessedDelta;
-            set => SetPropertyValue(nameof(LastProcessedDelta), ref lastProcessedDelta, value);
-        }
-
-
-        [Size(SizeAttribute.DefaultStringMappingFieldSize)]
-        public string LastPushedDelta
-        {
-            get => lastPushedDelta;
-            set => SetPropertyValue(nameof(LastPushedDelta), ref lastPushedDelta, value);
-        }
-        
-        public int LastTransactionLogProcessed
-        {
-            get => lastTransactionLogProcessed;
-            set => SetPropertyValue(nameof(LastTransactionLogProcessed), ref lastTransactionLogProcessed, value);
-        }
-    }
     public class XpoDelta : XPCustomObject, IDelta
     {
         public XpoDelta()
@@ -67,6 +21,7 @@ namespace BIT.Data.Sync.Xpo.DeltaStore
 
 
         Guid oid;
+   
         byte[] operation;
         string index;
         string identity;
@@ -74,12 +29,18 @@ namespace BIT.Data.Sync.Xpo.DeltaStore
         DateTime date;
 
 
+        //[Key(autoGenerate: true)]
+        //public Guid Oid
+        //{
+        //    get => oid;
+        //    set => SetPropertyValue(nameof(Oid), ref oid, value);
+        //}
 
-        [Key(autoGenerate: true)]
+        [Key(true)]
         public Guid Oid
         {
             get => oid;
-            private set => SetPropertyValue(nameof(Oid), ref oid, value);
+            set => SetPropertyValue(nameof(Oid), ref oid, value);
         }
         public DateTime Date
         {