+++ src/Tests/BIT.Data.Sync.Tests/SimpleDatabasesTest/SimpleDatabaseLocalTests.cs
@@ -29,7 +29,7 @@ namespace BIT.Data.Sync.Tests.SimpleDatabasesTest
             SimpleDatabase Master = new SimpleDatabase(MasterDeltaStore, "Master");
 
             //3 - Create a processor to allow the master to process other nodes deltas
-            SimpleDatabaseDeltaProcessor Master_DeltaProcessor = new SimpleDatabaseDeltaProcessor(Master.Data);
+            SimpleDatabaseDeltaProcessor Master_DeltaProcessor = new SimpleDatabaseDeltaProcessor(Master.Data, new MemorySequenceService(new YearSequencePrefixStrategy()));
 
             //4 - Create data and save it on the master
             SimpleDatabaseRecord Hello = new SimpleDatabaseRecord() { Key = Guid.NewGuid(), Text = "Hello" };
@@ -43,7 +43,7 @@ namespace BIT.Data.Sync.Tests.SimpleDatabasesTest
 
             //6 - Creating database A
             SimpleDatabase A_Database = new SimpleDatabase(A_DeltaStore, "A");
-            SimpleDatabaseDeltaProcessor A_DeltaProcessor = new SimpleDatabaseDeltaProcessor(A_Database.Data);
+            SimpleDatabaseDeltaProcessor A_DeltaProcessor = new SimpleDatabaseDeltaProcessor(A_Database.Data, new MemorySequenceService(new YearSequencePrefixStrategy()));
 
             //7 - Create data and save it on database A
             SimpleDatabaseRecord Hola = new SimpleDatabaseRecord() { Key = Guid.NewGuid(), Text = "Hola" };
@@ -57,7 +57,7 @@ namespace BIT.Data.Sync.Tests.SimpleDatabasesTest
 
             //9 - Creating database B
             SimpleDatabase B_Database = new SimpleDatabase(B_DeltaStore, "B");
-            SimpleDatabaseDeltaProcessor B_DeltaProcessor = new SimpleDatabaseDeltaProcessor(B_Database.Data);
+            SimpleDatabaseDeltaProcessor B_DeltaProcessor = new SimpleDatabaseDeltaProcessor(B_Database.Data, new MemorySequenceService(new YearSequencePrefixStrategy()));
 
             //10 - Create data and save it on database B
             SimpleDatabaseRecord Privet = new SimpleDatabaseRecord() { Key = Guid.NewGuid(), Text = "Privet" };